requirements.txt
discord.py
Pillow
praw
asyncio
python-dotenv
requests
discord
numpy
keras
nltk
pymongo
dnspython
bs4
praw
Pillow
wikipedia
